Deployed Web Version Updated to almost Final Version! We could submit this as is but I am still waiting for Shylens Edits on those SpashScreen and TitleScreen Images for the true Final Version.
Summary of Changes:
MiniGame#2 and it's corresponding Leaderboard added.
Fixed some issues with the Multiplayer Menu not showing some information correctly;
Mainly the Player's name instead of a hardcoded "Player 443".
And the Game Characters' Face Images not showing up correctly in that scene.
They show up now but the plugin is still in development so I used a
very roundabout way to have the FaceImages still show up, unfortunaly there is nothing
I can do about the pixelation of the original HighRes FaceImages.
We tried to have blockchain integration but there was just not enough time
to study and pursue it anymore in this hackathon. I know there's someone who
can make Web3 Games with RPGMakerMV though so it's not totally impossible for sure.